refactor: split backends/windows/_input.py into focused submixins (fixes #861)
_input.py had grown to 1,057 lines in a single InputMixin, bundling four
unrelated input domains. Split it into a focused package mirroring the
_shell/ precedent (#833), keeping InputMixin a thin composition so the
public WindowsBackend API is unchanged:

- _mouse.py (MouseMixin): click, scroll, drag, move_mouse
- _keyboard.py (KeyboardMixin): type_text, press_key, hotkey
- _uia_interact.py (UIAInteractMixin): UIA-pattern element interaction
  (click_element_uia, invoke_element, set/toggle/select/expand/focus,
  _init_comtypes_uia, _find_uia_element)
- _clipboard.py (ClipboardMixin): clipboard_get/set/clear/info

InputMixin(MouseMixin, KeyboardMixin, UIAInteractMixin, ClipboardMixin)
retains the app-list class constants (_SYSTEM_PROCESS_NAMES,
_UWP_HOST_PROCESS) consumed via self/backend by _shell/_app.py and
cli/_app/lifecycle.py. get_input_strategy is re-exported from the package
for backward compatibility.

Method bodies are moved verbatim — no behavior change. Tests that patch
get_input_strategy for keyboard methods are repointed to the new
_input._keyboard call site.
