+++ naturo/backends/windows/_input/_uia_interact.py
@@ -1,66 +1,22 @@
-"""Input handling: mouse, keyboard, clipboard, and UIA element interaction."""
+"""UIA element interaction: invoke, toggle, select, expand/collapse, focus.
+
+These methods drive UI elements through UI Automation patterns instead of
+synthetic ``SendInput`` events.  They are essential for UWP apps (hosted by
+ApplicationFrameHost.exe) where coordinate-based clicks do not reach inner
+content, and for schtasks / remote-session contexts where SendInput has no
+effect.
+"""
 
 from __future__ import annotations
 
 import logging
 from typing import Optional
 
-from naturo.backends.windows._strategies import get_input_strategy
-from naturo.errors import NaturoError
-
 logger = logging.getLogger(__name__)
 
 
-class InputMixin:
-    """Input handling: mouse, keyboard, clipboard, and UIA element interaction."""
-
-    # === Phase 2: Input ===
-
-    def click(self, x: Optional[int] = None, y: Optional[int] = None,
-              element_id: Optional[str] = None, button: str = "left",
-              double: bool = False, input_mode: str = "normal",
-              hwnd: Optional[int] = None) -> None:
-        """Click at coordinates or on a UI element.
-
-        If an element_id is provided, finds the element first and clicks its
-        center. If x/y coordinates are provided, moves to them first.
-        At least one of x/y or element_id must be given.
-
-        Args:
-            x: Target X coordinate. Required if element_id not given.
-            y: Target Y coordinate. Required if element_id not given.
-            element_id: Automation element ID to find and click.
-            button: Mouse button — "left", "right", or "middle".
-            double: True for double-click.
-            input_mode: Ignored for now (Phase 3 will add hardware/hook modes).
-            hwnd: Target window handle for element search.  When provided,
-                searches within this window instead of the foreground
-                window (#525).
-
-        Raises:
-            ValueError: If neither coordinates nor element_id is provided.
-            ElementNotFoundError: When element_id is given but not found.
-            NaturoCoreError: On system error.
-        """
-        core = self._ensure_core()
-        strategy = get_input_strategy(core, input_mode)
-
-        BUTTON_MAP = {"left": 0, "right": 1, "middle": 2}
-        btn = BUTTON_MAP.get(button.lower(), 0)
-
-        if element_id is not None:
-            # Find element and click its center
-            el = self.find_element(selector=element_id, hwnd=hwnd)
-            if el is None:
-                from naturo.errors import ElementNotFoundError
-                raise ElementNotFoundError(element_id)
-            cx = el.x + el.width // 2
-            cy = el.y + el.height // 2
-            strategy.click(cx, cy, btn, double)
-        elif x is not None and y is not None:
-            strategy.click(x, y, btn, double)
-        else:
-            raise ValueError("click: provide either (x, y) or element_id")
+class UIAInteractMixin:
+    """UIA element interaction via UI Automation patterns (no SendInput)."""
 
     def click_element_uia(
         self,
@@ -654,404 +610,3 @@ class InputMixin:
         except Exception as exc:
             logger.debug("UIA SetFocus unexpected error: %s", exc)
             return False
-
-    def type_text(self, text: str = "", delay_ms: int = 5, profile: str = "linear",
-                  wpm: int = 120, input_mode: str = "normal") -> None:
-        """Type text using SendInput.
-
-        Args:
-            text: UTF-8 string to type.
-            delay_ms: Delay between keystrokes (milliseconds). Default: 5.
-                For "human" profile, this is the base delay.
-            profile: "linear" for constant delay, "human" for variable speed.
-                     Human profile uses wpm to calculate delay.
-            wpm: Words per minute (used only when profile="human").
-            input_mode: Input method — "normal" (virtual key / Unicode) or
-                "hardware" (scan code / Phys32, bypasses game anti-cheat).
-
-        Raises:
-            NaturoCoreError: On system error.
-        """
-        import re
-
-        core = self._ensure_core()
-        strategy = get_input_strategy(core, input_mode)
-
-        actual_delay = delay_ms
-        if profile == "human" and wpm > 0:
-            # Average word = 5 chars, convert wpm to ms per char
-            ms_per_char = int(60_000 / (wpm * 5))
-            actual_delay = max(1, ms_per_char)
-
-        # (#840) SendInput's UNICODE path silently drops \n and \r
-        # control characters.  Split on line breaks and press Enter
-        # between segments so multiline text is typed correctly.
-        segments = re.split(r"\r\n|\r|\n", text)
-        for i, segment in enumerate(segments):
-            if segment:
-                strategy.type_text(segment, actual_delay)
-            if i < len(segments) - 1:
-                strategy.press_key("enter")
-
-    def press_key(self, key: str = "", input_mode: str = "normal") -> None:
-        """Press and release a named key.
-
-        Args:
-            key: Key name (enter, tab, escape, f1-f12, a-z, 0-9, etc.).
-            input_mode: Input method — "normal" (virtual key) or
-                "hardware" (scan code / Phys32).
-
-        Raises:
-            NaturoCoreError: If key name is unrecognized or on system error.
-        """
-        core = self._ensure_core()
-        strategy = get_input_strategy(core, input_mode)
-        strategy.press_key(key)
-
-    def hotkey(self, *keys: str, hold_duration_ms: int = 50,
-              input_mode: str = "normal") -> None:
-        """Press a hotkey combination.
-
-        Args:
-            *keys: Key names. Modifiers (ctrl, alt, shift, win) are recognized
-                   automatically. One non-modifier key is the base key.
-            hold_duration_ms: Not yet used.
-            input_mode: Input method — "normal" (virtual key) or
-                "hardware" (scan code / Phys32).
-
-        Example:
-            backend.hotkey("ctrl", "c")   # Copy
-            backend.hotkey("ctrl", "z")   # Undo
-
-        Raises:
-            NaturoCoreError: On invalid argument or system error.
-        """
-        core = self._ensure_core()
-        strategy = get_input_strategy(core, input_mode)
-        strategy.hotkey(*keys)
-
-    def scroll(self, direction: str = "down", amount: int = 3,
-               x: Optional[int] = None, y: Optional[int] = None,
-               smooth: bool = False) -> None:
-        """Scroll the mouse wheel.
-
-        Args:
-            direction: "up", "down", "left", or "right".
-            amount: Number of wheel notches (each = 120 delta units).
-            x: Move mouse to this X before scrolling. None = current position.
-            y: Move mouse to this Y before scrolling. None = current position.
-            smooth: Ignored (Phase 3 will add smooth scroll support).
-
-        Raises:
-            NaturoCoreError: On system error.
-        """
-        core = self._ensure_core()
-        strategy = get_input_strategy(core)
-
-        if x is not None and y is not None:
-            core.mouse_move(x, y)
-
-        WHEEL_DELTA = 120
-        horizontal = direction in ("left", "right")
-        # Up/right = positive, Down/left = negative
-        sign = 1 if direction in ("up", "right") else -1
-        delta = sign * amount * WHEEL_DELTA
-
-        strategy.scroll(delta, horizontal)
-
-    def drag(self, from_x: int = 0, from_y: int = 0, to_x: int = 0, to_y: int = 0,
-             duration_ms: int = 500, steps: int = 10,
-             trajectory: str = "linear", jitter: float = 0.0,
-             overshoot: float = 0.0, release_delay_ms: int = 0) -> None:
-        """Drag from one point to another.
-
-        Moves mouse to (from_x, from_y), holds left button, follows a
-        trajectory to (to_x, to_y), then releases the button.
-
-        Args:
-            from_x: Source X coordinate.
-            from_y: Source Y coordinate.
-            to_x: Destination X coordinate.
-            to_y: Destination Y coordinate.
-            duration_ms: Total drag duration in milliseconds.
-            steps: Number of intermediate move steps.
-            trajectory: Motion mode — ``"linear"`` (default), ``"bezier"``,
-                or ``"instant"``.
-            jitter: Max random perpendicular offset per step (pixels).
-            overshoot: Pixels to overshoot past target then correct back.
-            release_delay_ms: Pause in ms before releasing the button.
-
-        Raises:
-            NaturoCoreError: On system error.
-        """
-        import time
-        from naturo.backends.windows._trajectory import generate_trajectory
-
-        core = self._ensure_core()
-
-        points = generate_trajectory(
-            from_x, from_y, to_x, to_y,
-            mode=trajectory, duration_ms=duration_ms, steps=steps,
-            jitter=jitter, overshoot=overshoot,
-        )
-
-        core.mouse_move(from_x, from_y)
-        time.sleep(0.05)  # Brief settle before pressing
-        core.mouse_down(0)  # Press and hold left button
-
-        try:
-            for pt in points:
-                core.mouse_move(pt.x, pt.y)
-                if pt.delay_s > 0:
-                    time.sleep(pt.delay_s)
-            if release_delay_ms > 0:
-                time.sleep(release_delay_ms / 1000.0)
-        finally:
-            core.mouse_up(0)  # Always release, even on error
-
-    def move_mouse(self, x: int = 0, y: int = 0, *,
-                   trajectory: str = "instant",
-                   duration_ms: int = 500, steps: int | None = None,
-                   jitter: float = 0.0, overshoot: float = 0.0) -> None:
-        """Move the mouse cursor to absolute screen coordinates.
-
-        Args:
-            x: Target X coordinate.
-            y: Target Y coordinate.
-            trajectory: Motion mode — ``"instant"`` (default, teleport),
-                ``"linear"``, or ``"bezier"`` (human-like).
-            duration_ms: Total movement time in milliseconds (non-instant modes).
-            steps: Number of intermediate points (auto-calculated if omitted).
-            jitter: Max random perpendicular offset per step (pixels).
-            overshoot: Pixels to overshoot past target then correct back.
-
-        Raises:
-            NaturoCoreError: On system error.
-        """
-        import time
-        from naturo.backends.windows._trajectory import generate_trajectory
-
-        core = self._ensure_core()
-
-        if trajectory == "instant":
-            core.mouse_move(x, y)
-            return
-
-        # Get current cursor position as trajectory start
-        try:
-            import ctypes
-            from ctypes import wintypes
-            point = wintypes.POINT()
-            ctypes.windll.user32.GetCursorPos(ctypes.byref(point))  # type: ignore[union-attr]
-            start_x, start_y = point.x, point.y
-        except Exception:
-            # Non-Windows or error: just teleport
-            core.mouse_move(x, y)
-            return
-
-        points = generate_trajectory(
-            start_x, start_y, x, y,
-            mode=trajectory, duration_ms=duration_ms, steps=steps,
-            jitter=jitter, overshoot=overshoot,
-        )
-
-        for pt in points:
-            core.mouse_move(pt.x, pt.y)
-            if pt.delay_s > 0:
-                time.sleep(pt.delay_s)
-
-    def clipboard_get(self) -> str:
-        """Get text content from the clipboard.
-
-        Uses the pyperclip library as a portable clipboard interface.
-
-        Returns:
-            Clipboard text, or empty string if clipboard is empty.
-        """
-        try:
-            import pyperclip  # type: ignore
-            return pyperclip.paste() or ""
-        except ImportError:
-            # Fallback: use ctypes Win32 API directly
-            try:
-                import ctypes
-                import ctypes.wintypes
-                user32 = ctypes.windll.user32
-                kernel32 = ctypes.windll.kernel32
-                # Set proper restype/argtypes for 64-bit pointer safety
-                user32.GetClipboardData.restype = ctypes.c_void_p
-                kernel32.GlobalLock.restype = ctypes.c_void_p
-                kernel32.GlobalLock.argtypes = [ctypes.c_void_p]
-                kernel32.GlobalUnlock.argtypes = [ctypes.c_void_p]
-                if not user32.OpenClipboard(0):
-                    raise NaturoError("Failed to open clipboard")
-                try:
-                    CF_UNICODETEXT = 13
-                    h = user32.GetClipboardData(CF_UNICODETEXT)
-                    if not h:
-                        return ""
-                    ptr = kernel32.GlobalLock(h)
-                    if not ptr:
-                        raise NaturoError("Failed to lock clipboard memory")
-                    try:
-                        return ctypes.wstring_at(ptr)
-                    finally:
-                        kernel32.GlobalUnlock(h)
-                finally:
-                    user32.CloseClipboard()
-            except NaturoError:
-                raise
-            except Exception as exc:
-                raise NaturoError(f"Clipboard read failed: {exc}") from exc
-
-    def clipboard_set(self, text: str = "") -> None:
-        """Set the clipboard text content.
-
-        Args:
-            text: Text to place on the clipboard.
-        """
-        try:
-            import pyperclip  # type: ignore
-            pyperclip.copy(text)
-        except ImportError:
-            try:
-                import ctypes
-                import ctypes.wintypes
-                user32 = ctypes.windll.user32
-                kernel32 = ctypes.windll.kernel32
-                # Set proper restype/argtypes for 64-bit pointer safety
-                kernel32.GlobalAlloc.restype = ctypes.c_void_p
-                kernel32.GlobalAlloc.argtypes = [ctypes.c_uint, ctypes.c_size_t]
-                kernel32.GlobalLock.restype = ctypes.c_void_p
-                kernel32.GlobalLock.argtypes = [ctypes.c_void_p]
-                kernel32.GlobalUnlock.argtypes = [ctypes.c_void_p]
-                user32.SetClipboardData.restype = ctypes.c_void_p
-                user32.SetClipboardData.argtypes = [ctypes.c_uint, ctypes.c_void_p]
-                CF_UNICODETEXT = 13
-                GMEM_MOVEABLE = 2
-                encoded = (text + "\0").encode("utf-16-le")
-                h = kernel32.GlobalAlloc(GMEM_MOVEABLE, len(encoded))
-                if not h:
-                    raise NaturoError("Failed to allocate clipboard memory")
-                ptr = kernel32.GlobalLock(h)
-                if not ptr:
-                    kernel32.GlobalFree = kernel32.GlobalFree  # noqa: E731
-                    raise NaturoError("Failed to lock clipboard memory")
-                ctypes.memmove(ptr, encoded, len(encoded))
-                kernel32.GlobalUnlock(h)
-                if not user32.OpenClipboard(0):
-                    raise NaturoError("Failed to open clipboard")
-                try:
-                    user32.EmptyClipboard()
-                    user32.SetClipboardData(CF_UNICODETEXT, h)
-                finally:
-                    user32.CloseClipboard()
-            except NaturoError:
-                raise
-            except Exception as exc:
-                raise NaturoError(f"Clipboard write failed: {exc}") from exc
-
-    def clipboard_clear(self) -> None:
-        """Clear the clipboard contents."""
-        try:
-            import pyperclip  # type: ignore
-            pyperclip.copy("")
-        except ImportError:
-            try:
-                import ctypes
-                user32 = ctypes.windll.user32
-                if not user32.OpenClipboard(0):
-                    raise NaturoError("Failed to open clipboard")
-                try:
-                    user32.EmptyClipboard()
-                finally:
-                    user32.CloseClipboard()
-            except NaturoError:
-                raise
-            except Exception as exc:
-                raise NaturoError(f"Clipboard clear failed: {exc}") from exc
-
-    def clipboard_info(self) -> dict:
-        """Return information about the current clipboard contents.
-
-        Checks for text, image (CF_DIB/CF_BITMAP), and file (CF_HDROP)
-        formats using the Win32 API.
-
-        Returns:
-            Dictionary with format, size, has_text, has_image, has_files.
-        """
-        try:
-            import ctypes
-            user32 = ctypes.windll.user32
-            kernel32 = ctypes.windll.kernel32
-            user32.GetClipboardData.restype = ctypes.c_void_p
-            kernel32.GlobalLock.restype = ctypes.c_void_p
-            kernel32.GlobalLock.argtypes = [ctypes.c_void_p]
-            kernel32.GlobalUnlock.argtypes = [ctypes.c_void_p]
-            kernel32.GlobalSize.restype = ctypes.c_size_t
-            kernel32.GlobalSize.argtypes = [ctypes.c_void_p]
-
-            if not user32.OpenClipboard(0):
-                raise NaturoError("Failed to open clipboard")
-            try:
-                CF_UNICODETEXT = 13
-                CF_BITMAP = 2
-                CF_DIB = 8
-                CF_HDROP = 15
-
-                has_text = bool(user32.IsClipboardFormatAvailable(CF_UNICODETEXT))
-                has_image = bool(
-                    user32.IsClipboardFormatAvailable(CF_BITMAP)
-                    or user32.IsClipboardFormatAvailable(CF_DIB)
-                )
-                has_files = bool(user32.IsClipboardFormatAvailable(CF_HDROP))
-
-                size = 0
-                if has_text:
-                    fmt = "text"
-                    h = user32.GetClipboardData(CF_UNICODETEXT)
-                    if h:
-                        size = kernel32.GlobalSize(h)
-                elif has_image:
-                    fmt = "image"
-                    h = user32.GetClipboardData(CF_DIB)
-                    if h:
-                        size = kernel32.GlobalSize(h)
-                elif has_files:
-                    fmt = "files"
-                else:
-                    fmt = "empty"
-
-                return {
-                    "format": fmt,
-                    "size": size,
-                    "has_text": has_text,
-                    "has_image": has_image,
-                    "has_files": has_files,
-                }
-            finally:
-                user32.CloseClipboard()
-        except NaturoError:
-            raise
-        except Exception as exc:
-            raise NaturoError(f"Clipboard info failed: {exc}") from exc
-
-    # System/framework processes to exclude from app list — these have visible
-    # windows but are not user-facing applications.
-    _SYSTEM_PROCESS_NAMES: set[str] = {
-        "textinputhost.exe", "shellexperiencehost.exe",
-        "searchhost.exe", "startmenuexperiencehost.exe", "lockapp.exe",
-        "systemsettings.exe", "gamebar.exe", "gamebarftserver.exe",
-        "windowsinternal.composableshell.experiences.textinput.inputapp.exe",
-        "widgets.exe", "widgetservice.exe", "people.exe", "cortana.exe",
-        "secureinput.exe", "dwm.exe", "csrss.exe", "winlogon.exe",
-        "fontdrvhost.exe", "dllhost.exe", "sihost.exe", "ctfmon.exe",
-        "runtimebroker.exe", "backgroundtaskhost.exe", "taskhostw.exe",
-        "smartscreen.exe", "searchui.exe", "shellhost.exe",
-    }
-
-    # ApplicationFrameHost.exe hosts UWP apps (Calculator, Settings, etc.).
-    # Unlike other system processes, AFH windows with non-empty titles are
-    # real user-facing apps that should appear in app list.
-    _UWP_HOST_PROCESS: str = "applicationframehost.exe"
-
